[refactor] move the logger object from lib/util to lib/kuzzle
