+++ lib/core/statistics/index.js
@@ -24,10 +24,10 @@
 const errorsManager = require('../../util/errors').wrap('api', 'assert');
 
 /**
- * @class StatisticsController
+ * @class Statistics
  * @param {Kuzzle} kuzzle
  */
-class StatisticsController {
+class Statistics {
   constructor(kuzzle) {
     this.kuzzle = kuzzle;
 
@@ -312,4 +312,4 @@ function writeStats () {
   this.timer = setTimeout(() => writeStats.call(this), this.interval);
 }
 
-module.exports = StatisticsController;
+module.exports = Statistics;
