+++ lib/kuzzle/kuzzle.js
@@ -43,7 +43,7 @@ const InternalIndexBootstrap = require('../core/storage/bootstrap/internalIndexB
 const IndexStorage = require('../core/storage/indexStorage');
 const CacheEngine = require('../core/cache/cacheEngine');
 const StorageEngine = require('../core/storage/storageEngine');
-const Logger = require('../util/log');
+const Logger = require('./log');
 const signal = require('./signal');
 const vault = require('./vault');
 const shutdown = require('./shutdown');
