+++ lib/kuzzle/log.js
@@ -28,7 +28,7 @@ class Logger {
 
     this._useConsole();
 
-    this.eventEmitter.once('core:kuzzleStart', this._useLogger.bind(this));
+    this.eventEmitter.once('kuzzle:start', this._useLogger.bind(this));
   }
 
   _useConsole() {
